pyproject.toml
[tool.poetry]
name = "telegrambot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyTelegramBotAPI = "^4.2.2"
pytz = "^2021.3"
python-telegram-bot-calendar = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
